Aleksandrs K.: Lesson 12, rename level_5 to level_5_6
